backend/requirements.txt
langchain==0.3.0
langserve[all]==0.3.0
langgraph==0.2.34
langsmith>=0.1.126
langchain_openai>=0.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.7.4
openai==1.51.0
python-dotenv==1.0.0
websockets==12.0
numpy>=1.26.0,<2.0.0
pandas>=2.0.0,<3.0.0
dtw-python==1.3.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6
speechrecognition==3.10.0
pyttsx3==2.90
SQLAlchemy>=2.0.0,<3.0.0
requests>=2.32.0,<3.0.0
scipy>=1.15.0,<2.0.0